feat(activity): Updated Activity based on bids

+++ src/controllers/Bid.ts
@@ -4,6 +4,9 @@ import Bid from "../models/Bid";
 import ErrorHandler from "../providers/Error";
 import User from "../models/User";
 import ConnectedWallets from "../models/ConnectedWallets";
+import Activity from "../models/Activity";
+import ActivityController from "./Activity";
+
 class BidController{
     public static async createANewBid(req: Request, res: Response){
         try{
@@ -48,9 +51,23 @@ class BidController{
                     }
                 }
                 await foundBid.save()
+                let description: string=''
+                if(quantity){
+                    description=`${foundUser.name} updated their bid to ${amount} on ${foundNFT.name} for ${quantity} ${quantity>1? 'copies': 'copy'}`
+                }else{
+                    description=`${foundUser.name} updated their bid to ${amount} on ${foundNFT.name}`
+                }
+                await ActivityController.createActivityForGroups(description, res.locals.userId, nft_id)
                 return res.status(200).json({message: 'Bid Updated Success', bid: foundBid})  
             }else{
                 const bid=await Bid.create({amount, created_by: res.locals.userId, wallet_address, nft: nft_id, quantity: quantity})
+                let description: string=''
+                if(quantity){
+                    description=`${foundUser.name} created a bid of ${amount} on ${foundNFT.name} for ${quantity} ${quantity>1? 'copies': 'copy'}`
+                }else{
+                    description=`${foundUser.name} created a bid of ${amount} on ${foundNFT.name}`
+                }
+                await ActivityController.createActivityForGroups(description, res.locals.userId, nft_id)
                 return res.status(200).json({message: 'Bid Created Success', bid})  
             }
         }catch(err){
@@ -95,17 +112,31 @@ class BidController{
     public static async deleteBidById(req: Request, res: Response){
         try{
             const {bidId}=req.params
+            const foundUser=await User.findById(res.locals.userId)
+            if(!foundUser){
+                throw new Error('User Not Found')
+            }
+            console.log('Found User is', foundUser)
             const currentDate=new Date()
             currentDate.setMinutes(currentDate.getMinutes()-5)
+            const foundBid: any=await Bid.findById(bidId).populate('nft', 'name _id price sale_type content_hash token_address').lean()
+            if(foundBid.created_by.toString()!==res.locals.userId){
+                throw new Error('The Bid was not created by the logged in user')
+            }
+            if(!foundBid){
+                throw new Error('Bid Not Found')
+            }
             const response=await Bid.deleteOne({_id: bidId, created_by: res.locals.userId, createdAt: {'$lte': currentDate.toISOString()}})
             if(response.deletedCount<1){
                 throw new Error('Bid not found, or can not delete before 5 Minutes')
             }
+            await ActivityController.createActivityForGroups(`${foundUser.name} withdrew their bid on ${foundBid.nft.name}`, res.locals.userId, foundBid.nft?._id)
             return res.status(200).json({message: 'Bid Delete Success'})
         }catch(err){
             return ErrorHandler.APIErrorHandler(err, res)
         }
     }
+    //TODO: Verify, probably this method is redundant
     public static async editBidById(req: Request, res: Response){
         try{
             const {bidId}=req.params
